+++ elfitor/elfreader/ELFReader.h
@@ -146,7 +146,7 @@ public:
     Elf32_Rela* get_rela( Elf32_Shdr* shdr, int index = 0 );
     Elf32_Rel* get_rel( Elf32_Shdr* shdr, int index = 0 );
     Elf32_Shdr* va2section( size_t va );
-    char* get_sym_by_value( int value );
+    char* get_sym_by_value( size_t value );
 };
 
 class Elf64 : public File
@@ -173,5 +173,5 @@ public:
     Elf64_Rela* get_rela( Elf64_Shdr* shdr, int index = 0 );
     Elf64_Rel* get_rel( Elf64_Shdr* shdr, int index = 0 );
     Elf64_Shdr* va2section( size_t va );
-    char* get_sym_by_value( int value );
+    char* get_sym_by_value( size_t value );
 };
